Total efficiency calculation

+++ Strat.cpp
@@ -4,7 +4,7 @@
 #include <algorithm>
 #include <TGraph.h>
 #include "TMath.h"
-#include<TRandom.h>
+#include <TRandom3.h>
 #include "TGenPhaseSpace.h"
 #include "TLorentzVector.h"
 #include "TH1D.h"
@@ -104,7 +104,7 @@ void Strat()
 	double E = 0.0;
 	double sigma = 0.0;
 	double E_smear = 0.0;
-	for(int i = 0; i < Edep_small.size(); i++)
+	for(int i = 0; i < static_cast<int>(Edep_small.size()); i++)
 	{
 		E = Edep_small[i];
 		sigma = (E * 0.044)/(sqrt(E/1000));
